refactor: move command managers to separate directory
Improve codebase organization by separating command registration logic from
business logic services.

The following changes were made:

1. Relocated all *CommandManager.ts files from src/services to a new
   src/commands directory.
2. Updated relative imports in the moved files to maintain access to
   services and utilities.
3. Updated src/extension.ts to reflect the new file locations for
   extension activation.
4. Corrected test imports to ensure integration tests continue to pass.

This refactoring reduces the clutter in the services directory and
provides a clearer separation of concerns between logic and UI/Command
interaction layers.
