+++ src/commands/CommandManager.ts
@@ -1,14 +1,14 @@
 import * as vscode from 'vscode';
 import { Constants } from '../constants';
-import { FilterManager } from './FilterManager';
-import { HighlightService } from './HighlightService';
-import { ResultCountService } from './ResultCountService';
-import { LogProcessor } from './LogProcessor';
-import { Logger } from './Logger';
+import { FilterManager } from '../services/FilterManager';
+import { HighlightService } from '../services/HighlightService';
+import { ResultCountService } from '../services/ResultCountService';
+import { LogProcessor } from '../services/LogProcessor';
+import { Logger } from '../services/Logger';
 import { QuickAccessProvider } from '../views/QuickAccessProvider';
 import { FilterGroup, FilterItem } from '../models/Filter';
-import { JsonPrettyService } from './JsonPrettyService';
-import { SourceMapService } from './SourceMapService';
+import { JsonPrettyService } from '../services/JsonPrettyService';
+import { SourceMapService } from '../services/SourceMapService';
 import { FilterGroupCommandManager } from './FilterGroupCommandManager';
 import { FilterItemCommandManager } from './FilterItemCommandManager';
 import { FilterPropertyCommandManager } from './FilterPropertyCommandManager';
