+++ src/commands/NavigationCommandManager.ts
