+++ src/extension.ts
@@ -7,24 +7,24 @@ import { LogProcessor } from './services/LogProcessor';
 import { HighlightService } from './services/HighlightService';
 import { ResultCountService } from './services/ResultCountService';
 import { Logger } from './services/Logger';
-import { CommandManager } from './services/CommandManager';
+import { CommandManager } from './commands/CommandManager';
 import { AdbService } from './services/AdbService';
 import { AdbDeviceTreeProvider } from './views/AdbDeviceTreeProvider';
-import { AdbCommandManager } from './services/AdbCommandManager';
+import { AdbCommandManager } from './commands/AdbCommandManager';
 import { LogBookmarkService } from './services/LogBookmarkService';
 import { LogBookmarkWebviewProvider } from './views/LogBookmarkWebviewProvider';
-import { LogBookmarkCommandManager } from './services/LogBookmarkCommandManager';
+import { LogBookmarkCommandManager } from './commands/LogBookmarkCommandManager';
 import { JsonPrettyService } from './services/JsonPrettyService';
 import { JsonTreeWebview } from './views/JsonTreeWebview';
 import { SourceMapService } from './services/SourceMapService';
 import { FilteredLogDefinitionProvider } from './providers/FilteredLogDefinitionProvider';
 import { ShellCommanderService } from './services/ShellCommanderService';
 import { ShellCommanderTreeDataProvider } from './views/ShellCommanderTreeDataProvider';
-import { ShellCommanderCommandManager } from './services/ShellCommanderCommandManager';
+import { ShellCommanderCommandManager } from './commands/ShellCommanderCommandManager';
 import { Constants } from './constants';
 import { FilterGroup, FilterItem } from './models/Filter';
 import { FileHierarchyService } from './services/FileHierarchyService';
-import { NavigationCommandManager } from './services/NavigationCommandManager';
+import { NavigationCommandManager } from './commands/NavigationCommandManager';
 import { FileHierarchyLensProvider } from './providers/FileHierarchyLensProvider';
 
 let debounceTimer: NodeJS.Timeout | undefined;
